requirements.txt
bs4
datasets
python-docx
docx2txt
langchain
langchain-community
langchain-huggingface
langchain-openai
langgraph
lark
llama-index
matplotlib
numpy
ollama
olefile
openpyxl
pandas
pymupdf
pypdf
python-pptx
rank_bm25
rapidfuzz
streamlit
unstructured
weasyprint
